Refactor and cleanup the code
