+++ baremaps-core/pom.xml
@@ -84,6 +84,10 @@ limitations under the License.
     </dependency>
     <dependency>
       <groupId>org.apache.baremaps</groupId>
+      <artifactId>baremaps-dem</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.baremaps</groupId>
       <artifactId>baremaps-flatgeobuf</artifactId>
     </dependency>
     <dependency>
@@ -105,10 +109,6 @@ limitations under the License.
     <dependency>
       <groupId>org.apache.baremaps</groupId>
       <artifactId>baremaps-pmtiles</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.baremaps</groupId>
-      <artifactId>baremaps-raster</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.baremaps</groupId>
