+++ baremaps-core/src/main/java/org/apache/baremaps/tilestore/raster/VectorContourTileStore.java
@@ -24,12 +24,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.zip.GZIPOutputStream;
+import org.apache.baremaps.dem.ContourTracer;
+import org.apache.baremaps.dem.ElevationUtils;
 import org.apache.baremaps.maplibre.vectortile.Feature;
 import org.apache.baremaps.maplibre.vectortile.Layer;
 import org.apache.baremaps.maplibre.vectortile.Tile;
 import org.apache.baremaps.maplibre.vectortile.VectorTileEncoder;
-import org.apache.baremaps.raster.ContourTracer;
-import org.apache.baremaps.raster.ElevationUtils;
 import org.apache.baremaps.tilestore.TileCoord;
 import org.apache.baremaps.tilestore.TileStore;
 import org.apache.baremaps.tilestore.TileStoreException;
