+++ baremaps-dem/Dockerfile
