+++ baremaps-dem/src/test/resources/fuji.png
