+++ baremaps-server/src/main/resources/dem/favicon.ico
