+++ pom.xml
@@ -52,13 +52,13 @@ limitations under the License.
     <module>baremaps-cli</module>
     <module>baremaps-core</module>
     <module>baremaps-data</module>
+    <module>baremaps-dem</module>
     <module>baremaps-flatgeobuf</module>
     <module>baremaps-gdal</module>
     <module>baremaps-geoparquet</module>
     <module>baremaps-maplibre</module>
     <module>baremaps-openstreetmap</module>
     <module>baremaps-pmtiles</module>
-    <module>baremaps-raster</module>
     <module>baremaps-server</module>
     <module>baremaps-testing</module>
   </modules>
@@ -247,6 +247,11 @@ limitations under the License.
       </dependency>
       <dependency>
         <groupId>org.apache.baremaps</groupId>
+        <artifactId>baremaps-dem</artifactId>
+        <version>${project.parent.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.baremaps</groupId>
         <artifactId>baremaps-flatgeobuf</artifactId>
         <version>${project.parent.version}</version>
       </dependency>
@@ -273,11 +278,6 @@ limitations under the License.
       <dependency>
         <groupId>org.apache.baremaps</groupId>
         <artifactId>baremaps-pmtiles</artifactId>
-        <version>${project.parent.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.apache.baremaps</groupId>
-        <artifactId>baremaps-raster</artifactId>
         <version>${project.parent.version}</version>
       </dependency>
       <dependency>
